chg: dev: move get_duplicate func to main module
